+++ Source/ZEEngine/ZEGraphics/ZEDirect3D9/ZED3D9TextureCube.h
@@ -60,8 +60,8 @@ class ZED3D9TextureCube : public ZETextureCube, public ZED3D9ComponentBase
 
 		virtual ZEViewPort*				GetViewPort(ZETextureCubeFace Face);
 
-		virtual bool					Create(int EdgeLenght, ZETexturePixelFormat PixelFormat, bool RenderTarget = false);
-		virtual bool					Lock(ZETextureCubeFace Face, void** Buffer, unsigned int* Pitch);
+		virtual bool					Create(unsigned int EdgeLenght, ZETexturePixelFormat PixelFormat, bool RenderTarget = false);
+		virtual bool					Lock(ZETextureCubeFace Face, void** Buffer, size_t* Pitch);
 		virtual void					Unlock(ZETextureCubeFace Face);
 		virtual void					Release();
